+++ src/app/posts/new/_components/NewPostForm.tsx
@@ -9,7 +9,7 @@ import Image from "@tiptap/extension-image";
 import Link from "@tiptap/extension-link";
 import { EditorContent, useEditor } from "@tiptap/react";
 import StarterKit from "@tiptap/starter-kit";
-import { useCallback } from "react";
+import Menu from "./Menu";
 
 export default function NewPostForm() {
   const editor = useEditor({
@@ -38,13 +38,6 @@ export default function NewPostForm() {
     immediatelyRender: false,
   });
 
-  const setLink = useCallback(() => {
-    const url = window.prompt("URL を入力してください:");
-    if (url && editor) {
-      editor.chain().focus().setLink({ href: url }).run();
-    }
-  }, [editor]);
-
   if (!editor) {
     return null;
   }
@@ -69,77 +62,7 @@ export default function NewPostForm() {
           </TabsList>
           <TabsContent value="write">
             <Card>
-              <div className="border-b p-2 flex gap-2 flex-wrap">
-                <Button
-                  size="sm"
-                  variant={
-                    editor.isActive("heading", { level: 1 })
-                      ? "default"
-                      : "outline"
-                  }
-                  onClick={() =>
-                    editor.chain().focus().toggleHeading({ level: 1 }).run()
-                  }
-                >
-                  H1
-                </Button>
-                <Button
-                  size="sm"
-                  variant={
-                    editor.isActive("heading", { level: 2 })
-                      ? "default"
-                      : "outline"
-                  }
-                  onClick={() =>
-                    editor.chain().focus().toggleHeading({ level: 2 }).run()
-                  }
-                >
-                  H2
-                </Button>
-                <Button
-                  size="sm"
-                  variant={editor.isActive("bold") ? "default" : "outline"}
-                  onClick={() => editor.chain().focus().toggleBold().run()}
-                >
-                  太字
-                </Button>
-                <Button
-                  size="sm"
-                  variant={editor.isActive("italic") ? "default" : "outline"}
-                  onClick={() => editor.chain().focus().toggleItalic().run()}
-                >
-                  斜体
-                </Button>
-                <Button
-                  size="sm"
-                  variant={
-                    editor.isActive("bulletList") ? "default" : "outline"
-                  }
-                  onClick={() =>
-                    editor.chain().focus().toggleBulletList().run()
-                  }
-                >
-                  箇条書き
-                </Button>
-                <Button
-                  size="sm"
-                  variant={
-                    editor.isActive("orderedList") ? "default" : "outline"
-                  }
-                  onClick={() =>
-                    editor.chain().focus().toggleOrderedList().run()
-                  }
-                >
-                  番号付き
-                </Button>
-                <Button
-                  size="sm"
-                  variant={editor.isActive("link") ? "default" : "outline"}
-                  onClick={setLink}
-                >
-                  リンク
-                </Button>
-              </div>
+              <Menu editor={editor} />
               <EditorContent editor={editor} />
             </Card>
           </TabsContent>
